requirements.txt
google-cloud-bigquery==3.30.0
google-cloud-bigquery-storage==2.29.1
matplotlib==3.10.1
pandas==2.2.3
pandas-gbq==0.28.0
pytest==8.3.5
streamlit==1.43.2
streamlit-lottie==0.0.5
